requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pytz==2022.7.1
sqlparse==0.4.3
